refactored Empty text state, fixed empty state for parlay superstars, removed duplicated empty state styles
