+++ i18n/translations.en.json
@@ -150,7 +150,6 @@
     "The ticket was successfully submitted": "The ticket was successfully submitted",
     "There are currently no matches being played": "There are currently no matches being played",
     "There are currently no {{ticketsType}}": "There are currently no {{ticketsType}}",
-    "There are no superstars in current biweekly period": "There are no superstars in current biweekly period",
     "Tickets": "Tickets",
     "Tickets not redeemed within 90 days of market being resolved are forfeit.": "Tickets not redeemed within 90 days of market being resolved are forfeit.",
     "Total": "Total",
